refactor(ae): extract the build cache into its own TU (#1221)
Fourth cluster out of ae.c: the content-hashed build cache (key
computation, publish, stale-tmp GC, init) and the `ae cache` command move
to ae_cache.c. The fnv64 hashers and the lib-dir content walk stay static
to that file; the entry points the driver's run/build paths use
(compute_cache_key, cache_publish, remove_dsym_bundle, gc_stale_cache_tmp,
init_cache_dir) are declared in ae_internal.h.

Two symbols cross the new boundary and are exposed rather than duplicated:
the resolved cache directory s_cache_dir (defined in ae_cache.c, read by
the build paths) and tc_lib_dir_append (defined in ae.c, called by the
env-seeded lib-dir hashing). get_exe_path, which compute_cache_key folds
into the key, is now declared in the header too.

ae.c drops to 6,480 lines. Pure code motion, identical binary. Verified:
all nine tools TUs compile clean under clang and x86_64-w64-mingw32-gcc
with the real build flags, a full build works, `ae run` / `ae cache` /
`ae build` behave, and touching ae_cache.c recompiles only ae_cache.o.
