+++ CHANGELOG.md
@@ -19,7 +19,8 @@ next version number before tagging the release.
   constantly. Three command clusters moved into their own sources reached
   through a new `tools/ae_internal.h`: `ae_cross.c` (the zig cross-compile
   backend), `ae_version.c` (list/install/switch releases), and `ae_repl.c`
-  (the interactive REPL), taking `ae.c` to 6,842 lines. Pure code motion:
+  (the interactive REPL) plus `ae_cache.c` (content-hashed build cache,
+  publish, GC, and `ae cache`), taking `ae.c` to 6,480 lines. Pure code motion:
   `ae` is already a multi-TU link (`ae_help.c`, `ae_fmt.c`, `ae_bindgen.c`),
   and it spends its wall-clock in `zig cc` / `posix_run` / disk rather than
   its own driver code, so there is no runtime cost. Behavior is unchanged.
